+++ extensions/shortcut-library/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [Category Dropdown] - {PR_MERGE_DATE}
+
+- Pick an existing category from a searchable dropdown when adding or editing a shortcut
+- Create a new category from the form
+
 ## [Initial Release] - 2026-09-01
 
 - Browse and fuzzy-search your own custom keyboard shortcuts